Add test for rawIssuePreview and refactor
Upon adding a test for rawIssuePreview, I realized that the IssuePrint
interface was getting overloaded. Thus, I pulled the humanIssuePreview
and rawIssuePreview functionality out of it, scoping it down to just
formatting the outputs for the human readable issues and renaming it
to IssuePrintFormatter.

Notably, I think getAssigneeList, getLabelList, and getProjectList are all
in the wrong place... It seems to me like they may belong on api.Issue
itself, though that touches something way deeper in the codebase. They
will live on the IssuePrintFormatter interface for now, but I'll probably
want to move them somewhere else before long.