Refactor Issue and Label to handle their own formatted data
This moves getAssigneeList, getProjectList, and getLabelList out of the
IssuePrintFormatter and into either Issue or Label with a small amount of
refactoring to separate functionality that was previously intertwined.

- getAssigneeList -> GetAssigneeListString
- getProjectList  -> GetProjectListString
- getLabelList    -> SortAlphabeticallyIgnoreCase (Labels),
                     getColorizedLabelsList (IssuePrintFormatter)

GetAssigneeListString and GetProjectListString now have tests and can be
called from any Issue

getLabelList saw a few other changes and one design choice. Essentially,
the old getLabelList function was doing two things: sorting the labels
alphabetically, ignoring case, and colorizing the labels if a Color field
was included in the Label. I've separated this functionality into two
places: Labels now handle sorting of themselves with the
SortAlphabeticallyIgnoreCase method on the Labels struct, and
colorization is done by the IssuePrintFormatter.

This ultimately led to me removing colorization from the rawIssuePreview
labels, but can be added back in with a little thought and refactor if we
decide colorization of the rawIssuePreview is desireable.