Add parseCurrentBranchWithAtPushRef method to pr finder
This is a reimplementation of parseCurrentBranch, but leverages git's
@{push} rev-parse functionality to determine the correct selector rather
than rebuilding it from the BranchConfig.

Note, we cannot fully remove the BranchConfig yet, as we currently still
need the merge ref to see if we are set up to merge to a pr. In the
following commit, I will remove this dependency.